+++ R/utils-metadata.R
@@ -1,3 +1,22 @@
+# Modifications Copyright 2020 Justin Cally
+# Copyright 2018 Province of British Columbia
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# https://www.apache.org/licenses/LICENSE-2.0.txt
+#
+# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and limitations under the License.
+#
+# Modifications/State changes made to original work: 
+# + retained the specify_geom_name() and geom_col_name() but rewrote geom_col_name() to just look for 'gml:' string 
+# + feature_hits() does a similar job to bcdc_number_wfs_records() but has been rewritten to work with the Vicmap geoserver
+# + get_col_df() added and uses the DescribeFeatureType service to 
+
+
 #' The Number of Rows of the Promised Data
 #' 
 #' @description `feature_hits()` returns an integer of the number of rows that match the passed query/promise. 
@@ -10,9 +29,15 @@
 #' @export
 #'
 #' @examples
+#' \donttest{
 #' vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") %>%
 #'  feature_hits()
+#'  }
 feature_hits <- function(x) {
+  
+  if(!check_geoserver()) {
+    return(0)
+  }
 
   x$query$resultType <- "hits"
   x$query$outputFormat <- "text/xml"
@@ -20,6 +45,9 @@ feature_hits <- function(x) {
   
   request <- httr::build_url(x)
   response <- httr::GET(request)
+  
+  # stop if broken
+  httr::stop_for_status(response)
   
   parsed <- httr::content(response, encoding = "UTF-8")
   
@@ -40,11 +68,16 @@ feature_hits <- function(x) {
 #' @export
 #'
 #' @examples
+#' \donttest{
 #' # Return the name of the geometry column
 #' vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") %>% 
 #'   geom_col_name()
-#'   
+#'  }
 geom_col_name <- function(x) {
+  
+  if(!check_geoserver(timeout = 10, quiet = TRUE)) {
+    return(NULL)
+  }
   
   geom_col <- get_col_df(x) %>% 
     dplyr::filter(grepl(x = type, pattern = "gml:")) %>%
@@ -58,10 +91,11 @@ geom_col_name <- function(x) {
 #' @rdname geom_col_name
 #' @export
 #' @examples
+#' \donttest{
 #' # Return the column names as a character vector
 #' vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") %>% 
 #'   feature_cols()
-#'   
+#' }   
 feature_cols <- function(x) {
   
   return(get_col_df(x)$name)
@@ -84,16 +118,27 @@ specify_geom_name <- function(x, CQL_statement){
 #' @rdname geom_col_name  
 #' @export
 #' @examples
+#' \donttest{
 #' # Return a data.frame of the columns and their XML schema string datatypes
+#' try(
 #' vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") %>% 
 #'   get_col_df()
-#'   
+#'   )
+#'  }
 get_col_df <- function(x) {
+  
+  if(!check_geoserver(timeout = 10, quiet = TRUE)) {
+    return(NULL)
+  }
   
   layer <- x$query$version
   base_url_n_wfs <- substr(getOption("vicmap.base_url", default = base_wfs_url), start = 0, stop = nchar(getOption("vicmap.base_url", default = base_wfs_url)) - 3)
   
   r <- httr::GET(paste0(base_url_n_wfs, x$query$typeNames, "/wfs?service=wfs&version=", x$query$version, "&request=DescribeFeatureType"))
+  
+  # stop if broken
+  httr::stop_for_status(r)
+  
   c <- httr::content(r, encoding = "UTF-8", type="text/xml") 
   
   list <- xml2::xml_child(xml2::xml_child(xml2::xml_child(xml2::xml_child(c, "xsd:complexType"),