+++ R/utils-options.R
@@ -1,7 +1,26 @@
+# Modifications Copyright 2020 Justin Cally
+# Copyright 2019 Province of British Columbia
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# https://www.apache.org/licenses/LICENSE-2.0.txt
+#
+# Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and limitations under the License.
+#
+# Modifications/State changes made to original work: 
+# + Modified the options in vicmap_options() (from bcdc_options()) and changed their descriptions to be suited for the vicmap wfs
+# + Added vicmap.base_url as an options
+# + check_chunk_limit() rewritten for the vicmap WFS
+
+
 #' options
 #' 
 #' @description 
-#' #' This function retrieves bcdata specific options that can be set. These options can be set
+#' This function retrieves bcdata specific options that can be set. These options can be set
 #' using `option({name of the option} = {value of the option})`. The default options are purposefully
 #' set conservatively to hopefully ensure successful requests. Resetting these options may result in
 #' failed calls to the data catalogue. Options in R are reset every time R is re-started.
@@ -18,8 +37,10 @@
 #' 
 #' `vicmap.base_url` is the base wfs url used to query the geoserver.
 #'
-#' @return data.frame
+#' @return vicmap_options() returns a \code{data.frame}
 #' @export
+#' @examples 
+#' vicmap_options()
 vicmap_options <- function() {
   
   null_to_na <- function(x) {
@@ -29,7 +50,7 @@ vicmap_options <- function() {
   dplyr::tribble(
     ~ option, ~ value, ~default,
     "vicmap.max_geom_pred_size", null_to_na(getOption("vicmap.max_geom_pred_size")), as.character(4400),
-    "vicmap.chunk_limit",null_to_na(getOption("vicmap.chunk_limit")), as.character(70000),
+    "vicmap.chunk_limit",null_to_na(getOption("vicmap.chunk_limit")), as.character(1500),
     "vicmap.base_url", getOption("vicmap.base_url"), base_wfs_url
   )
 }
@@ -44,4 +65,3 @@ check_chunk_limit <- function(){
     stop(glue::glue("Your chunk value of {chunk_value} exceed the Vicmap Data Catalogue chunk limit"), call. = FALSE)
   }
 }
-